+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2022-01-20
+
 ### Fixed
 
 - Fix bug with NS delegation.
@@ -52,7 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Create api and ingress entries in Route53.
 
-[Unreleased]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/giantswarm/dns-operator-openstack/releases/tag/v0.1.0